test: revert the questionList approach
can't compile stampy.ts without CF env available
